Fix/back dashboard

+++ backend/src/main/java/capstone/offflow/Dashboard/Service/DashboardServiceImpl.java
@@ -11,6 +11,7 @@ import capstone.offflow.User.Domain.User;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,21 +24,20 @@ public class DashboardServiceImpl implements DashboardService{
     private final DashboardRepository dashboardRepository;
 
     @Override
-    public void createDashboard(DashboardDto dashboardDto, User user) {
+    public Dashboard createDashboard(DashboardDto dashboardDto, User user) {
         Dashboard dashboard = new Dashboard();
         dashboard.setDashboardName(dashboardDto.getDashboardName());
         dashboard.setStartDate(dashboardDto.getStartDate());
         dashboard.setUser(user);
 
-        if (dashboardDto.getMetadataDto() != null) {
-            DashboardMetadata metadata = MetadataDto.convertToEntity(dashboardDto.getMetadataDto());
+        if (dashboardDto.getMetadata() != null) {
+            DashboardMetadata metadata = MetadataDto.convertToEntity(dashboardDto.getMetadata());
             dashboard.setMetadata(metadata);
         } else {
             dashboard.setMetadata(null); // 명시적으로 null 처리 (선택사항)
         }
 
-        dashboardRepository.save(dashboard);
-        log.info("Dashboard 생성 완료 - {}", dashboard.getId());
+        return dashboardRepository.saveAndFlush(dashboard);
     }
 
     @Override
@@ -46,17 +46,21 @@ public class DashboardServiceImpl implements DashboardService{
         Dashboard dashboard = dashboardRepository.findByIdAndUser(id, user)
                 .orElseThrow(() -> new EntityNotFoundException("해당 id의 유저를 찾을 수 없습니다."));
 
-        //fetch join을 통해 섹션 + 상품까지 함꼐 조회한 경우
+        //fetch join을 통해 섹션까지 함꼐 조회한 경우
         return DashboardDto.convertToDto(dashboard);
     }
 
     @Override
     public void deleteDashboard(Long id, User user) {
-        Dashboard dashboard = dashboardRepository.findByIdAndUser(id, user)
+        Dashboard dashboard = dashboardRepository.findById(id) // 여기!! user 조건 빼기
                 .orElseThrow(() -> new EntityNotFoundException("삭제할 대시보드를 찾을 수 없습니다."));
 
-        dashboardRepository.delete(dashboard);
-        log.info("Dashboard 삭제완료 {}", dashboard.getId());
+        if (!dashboard.getUser().equals(user)) {
+            throw new AccessDeniedException("이 대시보드를 삭제할 권한이 없습니다.");
+        }
 
+        dashboardRepository.delete(dashboard);
+        log.info("Dashboard 삭제 완료: {}", dashboard.getId());
     }
+
 }
